# openapi.yaml
openapi: 3.0.4
info:
  title: Code RAG API
  description: >-
    API for managing indexed code projects and querying their vectorized source
    code using natural language. Projects support full CRUD via POST/GET/PUT/DELETE
    /projects. The service stores per-project code documents (functions, types, members, etc.)
    along with vector embeddings, and allows retrieving the most semantically similar
    pieces of code for a given natural language question via a pgvector-backed similarity search.


    All endpoints exclusively accept and return application/json. Client errors
    and server errors are reported using the RFC 7807 "Problem Details for HTTP APIs"
    format, with the exception of 404 responses, which are returned with no response body.
  version: 1.0.0
paths:
  /api/v1/projects/{projectId}/code-queries:
    post:
      tags:
        - Code Query
      summary: Query a project's indexed code using natural language
      description: >-
        Accepts a natural language question about a project's codebase, converts
        it into a vector embedding, and returns the code documents from that project whose
        stored embeddings are most semantically similar to the question.
                    
        When the project exists but no code document scores highly enough to be
        considered a match, the response is a 200 OK with an empty array - this is not
        treated as an error.

        A 404 is returned only when the projectId itself does not correspond to
        any project.
      parameters:
        - name: projectId
          in: path
          description: >-
            Identifier of the project to search, corresponding to the `id` field
            returned by

            `GET /projects`. Must be a positive 64-bit integer; any other format
            (e.g. a GUID or

            non-numeric string) results in a 400 response.
          required: true
          schema:
            type: string
      requestBody:
        description: The natural language question to search the project's code with.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeQueryRequest"
      responses:
        "200":
          description: >-
            The code documents most semantically similar to the natural language
            question, ordered by

            descending similarity. Returns an empty array when the project has
            no code documents, or

            none are found to be similar enough to the question.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CodeQueryResultResponse"
        "400":
          description: >-
            Either the projectId path parameter is not a valid positive integer
            (e.g. a GUID was

            supplied), or the request body is missing the question field, has an
            empty/blank question,

            or is otherwise malformed.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: >-
            No project exists with the given projectId. This is the only
            condition under which this

            endpoint returns 404; the response has no body.
        "500":
          description: >-
            An unhandled exception occurred while processing the request. This
            is the only condition

            under which this endpoint returns 500.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ServerErrorProblemDetails"
...
components:
  schemas:
    CodeQueryRequest:
      type: object
      properties:
        question:
          type: string
          description: >-
            A natural language ask describing the code the caller is looking for
            (e.g. "where is the

            retry logic for failed payments?"). This text is embedded and
            compared against the stored

            code documents' embeddings for the target project using cosine
            similarity. Must not be empty

            or blank.
          nullable: true
      additionalProperties: false
      description: Natural language search request scoped to a single project.
    CodeQueryResultResponse:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier of the matched code document.
          format: int64
        source_file:
          type: string
          description: Path to the source file this code document was extracted from,
            relative to the project root.
          nullable: true
        gitRawUrl:
          type: string
          description: Provides a hyperlink that redirects user to a raw view of file at it's git's page'
          nullable: true
        gitUrl:
          type: string
          description: Provides a hyperlink to project git repository.
        kind:
          type: string
          description: >-
            The kind of code element this document represents (e.g. `function`,
            `method`,

            `class`, `interface`) as determined by the analyzer that indexed the
            source.
          nullable: true
        type_name:
          type: string
          description: >-
            Name of the enclosing type (class, struct, interface, etc.), if any,
            that this document

            belongs to.
          nullable: true
        member:
          type: string
          description: >-
            Name of the specific member (method, property, field, function) this
            document represents, if

            applicable.
          nullable: true
        embedding_text:
          type: string
          description: >-
            The normalized text representation of this code element that was
            embedded and stored as a

            vector. Typically a signature, summary, or snippet describing the
            code element's purpose.
          nullable: true
        similarity:
          type: number
          description: >-
            Cosine similarity between the query's embedding vector and this
            document's stored embedding.

            Ranges from 1.0 (identical) to -1.0 (opposite); higher values
            indicate closer semantic matches.
          format: double
      additionalProperties: false
      description: >-
        A single code document matched against a natural language query,
        annotated with its computed

        similarity score. Serializes as snake_case.
    ExceptionDetails:
      type: object
      properties:
        exception_type:
          type: string
          description: Fully-qualified type name of the exception that was raised.
          nullable: true
        message:
          type: string
          description: The exception's message.
          nullable: true
        stack_trace:
          type: string
          description: The exception's captured stack trace.
          nullable: true
      additionalProperties: false
      description: Details of the unhandled exception that caused the request to fail.
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: {}
...
    ServerErrorProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
        exception:
          $ref: "#/components/schemas/ExceptionDetails"
      additionalProperties: {}
      description: >-
        Problem Details payload for 500 responses. Exposing the raised
        exception's details is a

        deliberate trade-off for ease of debugging in this API, per the OpenAPI
        contract, and should

        be disabled or redacted in a hardened deployment.
...